Microservicio Libros

+++ libros/src/main/java/dev/diemigo/libros/repository/LibroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LibroRepository extends JpaRepository<Libro,Integer> {
+public interface LibroRepository extends JpaRepository<Libro,Long> {
 }
